requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
whoosh==2.7.4
pdfminer.six==20231228
markdown==3.6
